+++ source/Messaging.Infrastructure/Configuration/CompositionRoot.cs
@@ -54,6 +54,7 @@ using Messaging.Infrastructure.Configuration.DataAccess;
 using Messaging.Infrastructure.Configuration.Processing;
 using Messaging.Infrastructure.Configuration.Serialization;
 using Messaging.Infrastructure.IncomingMessages;
+using Messaging.Infrastructure.IncomingMessages.RequestChangeOfSupplier;
 using Messaging.Infrastructure.MasterData.MarketEvaluationPoints;
 using Messaging.Infrastructure.OutgoingMessages;
 using Messaging.Infrastructure.OutgoingMessages.AccountingPointCharacteristics;
